+++ app/loja/eventos/page.tsx
@@ -2,12 +2,16 @@
 
 import { EventForm } from '@/components/organisms'
 
-export default function EventosFormPage() {
+export default function EventosFormPage({
+  searchParams,
+}: {
+  searchParams: { evento?: string }
+}) {
   return (
     <main className="px-4 py-10 flex justify-center">
       <div className="w-full max-w-xl space-y-6">
         <h1 className="text-3xl font-bold text-center">Inscrição em Evento</h1>
-        <EventForm />
+        {searchParams.evento && <EventForm eventoId={searchParams.evento} />}
       </div>
     </main>
   )
